YUK-6871 Utility Pro Thermostat - additional cleanup
- Removed legacy thermostat schedule classes including: dbpersistents, lites and serializable
- Reworked the way we handle energy company default schedules

+++ yukon-client/stars/src/com/cannontech/stars/dr/advice/StarsDaoAdvice.java
@@ -49,18 +49,19 @@ public class StarsDaoAdvice {
 
     }
 
-    @After("bean(thermostatScheduleDao) && saveMethodNamePointCut() && args(schedule)")
-    public void doThermostatScheduleAction(ThermostatSchedule schedule)
-            throws Throwable {
+    @After("bean(thermostatScheduleDao) && saveMethodNamePointCut() && args(schedule, energyCompany)")
+    public void doThermostatScheduleAction(ThermostatSchedule schedule, 
+    		LiteStarsEnergyCompany energyCompany)
+    	throws Throwable {
 
         Integer accountId = schedule.getAccountId();
-        LiteStarsEnergyCompany energyCompany = mappingDao.getCustomerAccountEC(accountId);
-
-        // Clear and reload account info cache entry
-        LiteStarsCustAccountInformation custAccountInformation = energyCompany.getCustAccountInformation(accountId,
-                                                                                                         false);
-        energyCompany.deleteCustAccountInformation(custAccountInformation);
-
+        if(accountId != null) {
+	
+	        // Clear and reload account info cache entry
+	        LiteStarsCustAccountInformation custAccountInformation = energyCompany.getCustAccountInformation(accountId,
+	                                                                                                         false);
+	        energyCompany.deleteCustAccountInformation(custAccountInformation);
+        }
 
     }
